+++ test/unit/fuzz-tools.test.js
@@ -512,6 +512,7 @@ import * as ridleyJoplingV534 from '../../lib/ridley-jopling-v534.js';
 import * as capthusV535 from '../../lib/capthus-v535.js';
 import * as hardmanV536 from '../../lib/hardman-v536.js';
 import * as alsfrsRV537 from '../../lib/alsfrs-r-v537.js';
+import * as neosV538 from '../../lib/neos-v538.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -969,6 +970,7 @@ const MODULES = {
   'capthus-v535.js': capthusV535,
   'hardman-v536.js': hardmanV536,
   'alsfrs-r-v537.js': alsfrsRV537,
+  'neos-v538.js': neosV538,
   'tb-testing.js': tbTesting,
 };